Rename hl to hierarchical_label
